Respect semantics of headers
For formatting reasons, the title of a post was listed as ## in
the markdown, which is really a second-level header. This made a
bit of sense since the the only other headers were the categories
and those were top-level. However, other second level headers
have been used, breaking the technicality.

The new template respects header context and is, I believe, better
looking anyway.